add function to check if payment was done and regist the donation in streamlabs

+++ src/models/DonoRequestBody.go
@@ -1,11 +1,11 @@
 package models
 
 type DonoRequestBody struct {
-	Amount        int    `json:"amount"`
-	UserName      string `json:"userName"`
-	Message       string `json:"message"`
-	ImgUrl        string `json:"imgUrl"`
-	SuccessDomain string `json:"successDomain"`
-	FailDomain    string `json:"failDomain"`
-	ChannelId     int    `json:"channelId"`
+	Amount        float32 `json:"amount"`
+	UserName      string  `json:"userName"`
+	Message       string  `json:"message"`
+	ImgUrl        string  `json:"imgUrl"`
+	SuccessDomain string  `json:"successDomain"`
+	FailDomain    string  `json:"failDomain"`
+	ChannelId     int     `json:"channelId"`
 }
